Working calculation functions
Tested on edit page
