+++ app/imports/api/spreadsheet/BudgetPLCollection.js
@@ -4,145 +4,8 @@ import { Roles } from 'meteor/alanning:roles';
 import BaseCollection from '../base/BaseCollection';
 import { ROLE } from '../role/Role';
 import { BudgetPLSchema } from '../schemas/BudgetPLSchema';
-import { calcFringeBenefit, calcFringeBenefitManagement, calcSalary, sumExpenses, sumFringe, sumFringeBenefits, sumSurplus, sumTotalRevenue } from './BudgetPlFunctions';
+import { calcFringeBenefit, calcFringeBenefitManagement, calcSalary, getYearPercentages, sumExpenses, sumFringe, sumFringeBenefits, sumSurplus, sumTotalRevenue } from './BudgetPLFunctions';
 
-const percentages = [
-  {
-    year: 2,
-    percentages: {
-      pension_accumulation: 15.00,
-      retiree_health_insurance: 7.96,
-      other_post_employment_benefits: 0.00,
-      employees_health_fund: 7.02,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.22,
-      unemployment_compensation: 0.91,
-      pension_administration: 0.00,
-      composite_rate: 39.76,
-    },
-  },
-  {
-    year: 3,
-    percentages: {
-      pension_accumulation: 15.50,
-      retiree_health_insurance: 10.35,
-      other_post_employment_benefits: 0.00,
-      employees_health_fund: 6.84,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 0.88,
-      unemployment_compensation: 0.31,
-      pension_administration: 0.01,
-      composite_rate: 41.54,
-    },
-  },
-  {
-    year: 4,
-    percentages: {
-      pension_accumulation: 16.00,
-      retiree_health_insurance: 10.12,
-      other_post_employment_benefits: 0.00,
-      employees_health_fund: 6.81,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.16,
-      unemployment_compensation: 0.25,
-      pension_administration: 0.00,
-      composite_rate: 41.99,
-    },
-  },
-  {
-    year: 5,
-    percentages: {
-      pension_accumulation: 16.50,
-      retiree_health_insurance: 10.12,
-      other_post_employment_benefits: 0.00,
-      employees_health_fund: 6.81,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.16,
-      unemployment_compensation: 0.25,
-      pension_administration: 0.00,
-      composite_rate: 42.49,
-    },
-  },
-  {
-    year: 6,
-    percentages: {
-      pension_accumulation: 17.00,
-      retiree_health_insurance: 8.07,
-      other_post_employment_benefits: 7.78,
-      employees_health_fund: 7.62,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.27,
-      unemployment_compensation: 0.15,
-      pension_administration: 0.00,
-      composite_rate: 49.54,
-    },
-  },
-  {
-    year: 7,
-    percentages: {
-      pension_accumulation: 17.00,
-      retiree_health_insurance: 9.39,
-      other_post_employment_benefits: 12.69,
-      employees_health_fund: 7.60,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.06,
-      unemployment_compensation: 0.09,
-      pension_administration: 0.01,
-      composite_rate: 55.48,
-    },
-  },
-  {
-    year: 8,
-    percentages: {
-      pension_accumulation: 18.00,
-      retiree_health_insurance: 10.14,
-      other_post_employment_benefits: 14.33,
-      employees_health_fund: 7.69,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.24,
-      unemployment_compensation: 0.02,
-      pension_administration: 0.01,
-      composite_rate: 59.08,
-    },
-  },
-  {
-    year: 9,
-    percentages: {
-      pension_accumulation: 19.00,
-      retiree_health_insurance: 10.14,
-      other_post_employment_benefits: 14.33,
-      employees_health_fund: 7.69,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.24,
-      unemployment_compensation: 0.02,
-      pension_administration: 0.01,
-      composite_rate: 60.08,
-    },
-  },
-  {
-    year: 10,
-    percentages: {
-      pension_accumulation: 22.00,
-      retiree_health_insurance: 10.14,
-      other_post_employment_benefits: 14.33,
-      employees_health_fund: 7.69,
-      social_security: 6.20,
-      medicare: 1.45,
-      workers_compensation: 1.24,
-      unemployment_compensation: 0.02,
-      pension_administration: 0.01,
-      composite_rate: 63.08,
-    },
-  },
-];
 export const budgetPLPublications = {
   budgetPL: 'budgetPL',
   budgetPLAdmin: 'budgetPLAdmin',
@@ -153,37 +16,33 @@ class BudgetPLCollection extends BaseCollection {
   }
 
   calculateValues({ year, Revenue, Expenses, ExpenditurePerAudited }) {
-    const yearPercentages = percentages.find((entry) => entry.year === year);
+    const yearPercentages = getYearPercentages(year);
     const summedRevenue = sumTotalRevenue(Revenue);
-    const summedExpenses = sumExpenses(summedRevenue, Expenses);
+    const summedExpenses = sumExpenses(Expenses);
     const summedSurplus = sumSurplus(summedRevenue, summedExpenses);
 
     // todo: Rename FringeAdminManagement to PersonnelFringeManagement
     const personnelFringeManagement = Expenses.FringeAdminManagement;
-    const personnelFringeStaff = Expenses.PersonnelFringeStaff;
+    const personnelFringeStaff = Expenses.PersonnelFringeAdminStaff;
     const personnelFringeAdmin = Expenses.PersonnelFringeAdmin;
-
     const fringeBenefitsManagement = calcFringeBenefitManagement(year, yearPercentages, personnelFringeManagement.FringeBenefits, personnelFringeManagement.salary);
-
     const summedFringeBenefitManagement = sumFringeBenefits(fringeBenefitsManagement);
+    const summedFringe = sumFringe(summedFringeBenefitManagement, personnelFringeManagement.salary, ExpenditurePerAudited.management, Expenses.personnel);
 
-    const { summedFringeManagement, summedFringeStaff, summedFringeAdmin } = sumFringe(summedFringeBenefitManagement, personnelFringeManagement.salary, ExpenditurePerAudited.management);
-
-    const fringeBenefitsStaff = calcFringeBenefit(year, yearPercentages, personnelFringeStaff.FringeBenefits, summedFringeStaff);
-    const fringeBenefitsAdmin = calcFringeBenefit(year, yearPercentages, personnelFringeAdmin.FringeBenefits, summedFringeAdmin);
+    const fringeBenefitsStaff = calcFringeBenefit(year, yearPercentages, personnelFringeStaff.FringeBenefits, summedFringe.fringeStaff);
+    const fringeBenefitsAdmin = calcFringeBenefit(year, yearPercentages, personnelFringeAdmin.FringeBenefits, summedFringe.fringeAdmin);
 
     const summedFringeBenefitStaff = sumFringeBenefits(fringeBenefitsStaff);
     const summedFringeBenefitAdmin = sumFringeBenefits(fringeBenefitsAdmin);
-
-    const salaryStaff = calcSalary(yearPercentages.composite_rate, summedFringeStaff);
-    const salaryAdmin = calcSalary(yearPercentages.composite_rate, summedFringeAdmin);
-
+    const salaryStaff = calcSalary(yearPercentages.composite_rate, summedFringe.fringeStaff);
+    const salaryAdmin = calcSalary(yearPercentages.composite_rate, summedFringe.fringeAdmin);
     const updatedFringeManagement = {
+      salary: personnelFringeManagement.salary,
       FringeBenefits: {
         ...fringeBenefitsManagement,
         fringeBenefitsSum: summedFringeBenefitManagement,
       },
-      personnelFringeSum: summedFringeManagement,
+      personnelFringeSum: summedFringe.fringeManagement,
     };
     const updatedFringeStaff = {
       salary: salaryStaff,
@@ -191,7 +50,7 @@ class BudgetPLCollection extends BaseCollection {
         ...fringeBenefitsStaff,
         fringeBenefitsSum: summedFringeBenefitStaff,
       },
-      personnelFringeSum: summedFringeStaff,
+      personnelFringeSum: summedFringe.fringeStaff,
     };
     const updatedFringeAdmin = {
       salary: salaryAdmin,
@@ -199,7 +58,7 @@ class BudgetPLCollection extends BaseCollection {
         ...fringeBenefitsAdmin,
         fringeBenefitsSum: summedFringeBenefitAdmin,
       },
-      personnelFringeSum: summedFringeAdmin,
+      personnelFringeSum: summedFringe.fringeAdmin,
     };
     const updatedRevenue = {
       ...Revenue,
@@ -212,7 +71,7 @@ class BudgetPLCollection extends BaseCollection {
       FringeAdminManagement: updatedFringeManagement,
       totalExpenses: summedExpenses,
     };
-    return { updatedRevenue, updatedExpenses, summedSurplus, ExpenditurePerAudited };
+    return { updatedRevenue, updatedExpenses, summedSurplus };
   }
 
   /**
@@ -249,14 +108,13 @@ class BudgetPLCollection extends BaseCollection {
    * @param ExpenditurePerAudited Sub document for expenditure line items per audited financials values.
    * @return {String} the docID of the new document.
    */
-  update(docID, { year, Revenue, Expenses, ExpenditurePerAudited }) {
-    const beforeCalc = { year, Revenue, Expenses, ExpenditurePerAudited };
+  update(docID, { year, Revenue, Expenses, surplus, ExpenditurePerAudited }) {
     try {
       if (!docID) {
         throw new Meteor.Error('invalid-argument', 'Missing crucial field for updating BudgetPLCollection');
       }
-      const updateData = this.calculateValues(beforeCalc);
-
+      const calcData = this.calculateValues({ year, Revenue, Expenses, ExpenditurePerAudited });
+      const updateData = { Revenue: calcData.updatedRevenue, Expenses: calcData.updatedExpenses, surplus: calcData.summedSurplus, ExpenditurePerAudited };
       this._collection.update(docID, { $set: updateData });
     } catch (e) {
       throw new Meteor.Error('BudgetPLCollection.update', e.message || 'Update BudgetPLCollection failed.');
